requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
jinja2==3.1.6
python-multipart==0.0.20
bcrypt==4.0.1
python-jose[cryptography]==3.5.0
pytest==8.3.5
httpx==0.28.1
python-dotenv==1.1.0
